refactor(shared): moved to standalone sub-project
