+++ api-legacy/src/main/java/ga/wawupc/api/MappingConfiguration.java
@@ -1,7 +1,9 @@
-package ga.wawupc.api.shared.mapping;
+package ga.wawupc.api;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import ga.wawupc.api.shared.mapping.EnhancedModelMapper;
 
 @Configuration("enhancedModelMapperConfiguration")
 public class MappingConfiguration {
