+++ api-shared/src/main/java/ga/wawupc/api/shared/domain/service/communication/BaseResponse.java
