+++ api-shared/src/main/java/ga/wawupc/api/shared/exception/ResourceNotFoundException.java
