+++ api-shared/src/test/java/ga/wawupc/api/shared/JsonUtil.java
@@ -1,4 +1,4 @@
-package ga.wawupc.api;
+package ga.wawupc.api.shared;
 
 import java.io.IOException;
